Make source and dest config a dict
The source and destination fields are now a dict
that has the options path and mount.

Example:

- source:
    path: /some/source
  destination:
    path: /some/destination

This is to support multiple options that are
specific to the source or destination.

This patch also adds the mount option that will
validate that the source and/or destination is
an actual mount, example:

- source:
    path: /some/source
    mount: yes
  destination:
    path: /some/destination
    mount: no

This will validate that /some/source is a mount
but /some/destination is not before starting.